+++ client/src/scripts/others/worker_server.ts
@@ -5,12 +5,12 @@ import { ConfigType } from "common/scripts/config/config.ts";
 import { WorkerSocket } from "common/scripts/engine/server_offline/worker_socket.ts";
 import { SimpleBotAi } from "../../../../server/src/game_server/player/simple_bot_ai.ts";
 import { Backpacks } from "common/scripts/definitions/items/backpacks.ts";
-import { Armors } from "common/scripts/definitions/items/equipaments.ts";
 import { Consumibles } from "common/scripts/definitions/items/consumibles.ts";
 import { type GameItem } from "common/scripts/definitions/utils.ts";
 import { Melees } from "common/scripts/definitions/items/melees.ts";
 import { Boosts, BoostType } from "common/scripts/definitions/player/boosts.ts";
 import { Skins } from "common/scripts/definitions/loadout/skins.ts";
+import { Helmets, Vests } from "common/scripts/definitions/items/equipaments.ts";
 self.onerror = (e) => {
     console.error("Worker error:", e,e.valueOf())
 };
@@ -79,8 +79,8 @@ self.onmessage = (ev) => {
         if(Math.random()<=0.5){
             bot.inventory.set_weapon(0,Melees.getFromString(random.choose(["hammer","axe"])))
         }
-        bot.vest=Armors.getFromString("tactical_vest")
-        bot.helmet=Armors.getFromString("tactical_helmet")
+        bot.vest=Vests.getFromString("tactical_vest")
+        bot.helmet=Helmets.getFromString("tactical_helmet")
         bot.inventory.give_item(Consumibles.getFromString("medikit") as unknown as GameItem,4)
         bot.inventory.give_item(Consumibles.getFromString("yellow_pills") as unknown as GameItem,2)
         bot.inventory.give_item(Consumibles.getFromString("red_pills") as unknown as GameItem,2)