+++ common/scripts/engine/geometry.ts
@@ -1080,15 +1080,8 @@ export const Collision=Object.freeze({
         return ccw(p1, q1, q2) !== ccw(p2, q1, q2) &&
             ccw(p1, p2, q1) !== ccw(p1, p2, q2);
     },
-    point_on_segment(p: Vec2, a: Vec2, b: Vec2, epsilon = 1e-6): boolean {
-        const cross = (b.y - a.y) * (p.x - a.x) - (b.x - a.x) * (p.y - a.y);
-        if (Math.abs(cross) > epsilon) return false;
-
-        const dot = (p.x - a.x) * (b.x - a.x) + (p.y - a.y) * (b.y - a.y);
-        if (dot < 0) return false;
-
-        const lenSq = (b.x - a.x) ** 2 + (b.y - a.y) ** 2;
-        return dot <= lenSq;
-    }
-
 })
+export interface Rect{
+    min:Vec2
+    max:Vec2
+}