+++ server/src/game_server/others/modeManager.ts
@@ -176,14 +176,14 @@ export class SoloGamemodeManager extends GamemodeManager{
                         defs:[
                             {
                                 rivers:[
-                                    {sub_river_width:2,width:10,width_variation:1,sub_river_chance:0.5},
-                                    {sub_river_width:1,width:15,width_variation:1,sub_river_chance:0.1},
+                                    {sub_river_width:2,width:7,width_variation:1,sub_river_chance:0.5},
+                                    {sub_river_width:1,width:8,width_variation:1,sub_river_chance:0.1},
                                 ],
                                 weight:10
                             },
                             {
                                 rivers:[
-                                    {sub_river_width:3,width:20,width_variation:1,sub_river_chance:0.9},
+                                    {sub_river_width:3,width:15,width_variation:1,sub_river_chance:0.9},
                                 ],
                                 weight:1
                             }